Add support for local branch pruning in mirrors
When updating mirrors, set fetch.prune=true such that the
mirrors update do not fail due to conflicting local branches.

Note that while allowing a safe mirror update this does not
fix the problem with repo that does not manage corectly
local branches conflict in the users's own repo dir.

Actually repo >= 1.12.21 support local pruning on sync in
the case where a file must be changed to a directory in the
references tree.
Though this version of repo still does not support pruning for
the case where a directory must be changed to a file in the
reference tree.

Change-Id: Ic8a864fa2f5262f526cc3146a552117fbbf62184